+++ src/main/java/com/onepiece/otboo/infra/security/config/ProdSecurityConfig.java
@@ -1,5 +1,7 @@
 package com.onepiece.otboo.infra.security.config;
 
+import com.onepiece.otboo.infra.security.config.factory.SecurityFilterChainFactory;
+import com.onepiece.otboo.infra.security.config.helper.SecurityAuthorizeRequestHelper;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
