+++ src/test/java/com/onepiece/otboo/domain/auth/controller/AuthControllerTest.java
@@ -13,8 +13,8 @@ import com.onepiece.otboo.domain.user.entity.User;
 import com.onepiece.otboo.domain.user.exception.UserNotFoundException;
 import com.onepiece.otboo.domain.user.repository.UserRepository;
 import com.onepiece.otboo.infra.api.mail.service.MailService;
-import com.onepiece.otboo.infra.security.config.TestSecurityConfig;
 import com.onepiece.otboo.infra.security.jwt.JwtProvider;
+import com.onepiece.otboo.infra.security.testconfig.TestSecurityConfig;
 import jakarta.servlet.http.Cookie;
 import java.time.Instant;
 import java.util.Optional;
